+++ packages/@storefront/core/src/services/pastPurchases.ts
@@ -1,6 +1,8 @@
-import { Events, Routes } from '@storefront/flux-capacitor';
+import { Events, Routes, Store } from '@storefront/flux-capacitor';
 import { core, BaseService } from '../core/service';
+import UrlBeautifier from '../core/url-beautifier';
 import StoreFront from '../storefront';
+import Utils from './urlUtils';
 
 class PastPurchaseService extends BaseService<PastPurchaseService.Options> {
 
@@ -8,15 +10,25 @@ class PastPurchaseService extends BaseService<PastPurchaseService.Options> {
     super(app, opts);
 
     this.app.flux.on(Events.AUTOCOMPLETE_QUERY_UPDATED, () => this.app.flux.actions.receiveSaytPastPurchases([]));
-    this.app.flux.on(Events.PAST_PURCHASE_CHANGED, this.fetchProducts, this);
+    this.app.flux.on(Events.PAST_PURCHASE_CHANGED, this.pushState, this);
+    this.app.flux.on(Events.PAST_PURCHASE_URL_UPDATED, this.fetchProducts, this);
   }
 
   init() {
     this.app.flux.store.dispatch(this.app.flux.actions.fetchPastPurchases());
   }
 
-  fetchProducts() {
-    this.app.flux.saveState(Routes.PAST_PURCHASE);
+  pushState() {
+    this.app.flux.pushState({ route: Routes.PAST_PURCHASE });
+  }
+
+  fetchProducts(urlState: Store.History) {
+    const state = this.app.flux.store.getState();
+    const request = Utils.pastPurchaseStateToRequest(<UrlBeautifier.SearchUrlState>urlState.request, state);
+    const newState = Utils.mergePastPurchaseState(state, <UrlBeautifier.SearchUrlState>urlState.request);
+
+    this.app.flux.refreshState(newState);
+    this.app.flux.store.dispatch(this.app.flux.actions.fetchPastPurchaseProducts({ request }));
   }
 }