+++ packages/@storefront/flux-capacitor/src/core/reducers/data/index.ts
@@ -4,6 +4,7 @@ import area from './area';
 import autocomplete from './autocomplete';
 import collections from './collections';
 import details from './details';
+import history from './history';
 import infiniteScroll from './infinite-scroll';
 import navigations from './navigations';
 import page from './page';
@@ -24,6 +25,7 @@ export default redux.combineReducers({
   collections,
   details,
   fields: (state = []) => state,
+  history,
   infiniteScroll,
   navigations,
   page,