+++ packages/@storefront/flux-capacitor/src/core/reducers/is-fetching.ts
@@ -43,7 +43,7 @@ export default function updateIsFetching(state: State = DEFAULT_FETCHING, action
     case Actions.RECEIVE_AUTOCOMPLETE_SUGGESTIONS: return doneFetching(state, 'autocompleteSuggestions');
     case Actions.FETCH_AUTOCOMPLETE_PRODUCTS: return startFetching(state, 'autocompleteProducts');
     case Actions.RECEIVE_AUTOCOMPLETE_PRODUCTS: return doneFetching(state, 'autocompleteProducts');
-    case Actions.FETCH_PRODUCT_DETAILS: return action.payload.redirect
+    case Actions.FETCH_PRODUCT_DETAILS: return action.payload.buildAndParse
       ? { ...startFetching(state, 'details'), search: false }
       : startFetching(state, 'details');
     case Actions.RECEIVE_DETAILS: return doneFetching(state, 'details');